+++ src/main/java/com/example/demo/entity/Sensor.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,7 +16,7 @@ public class Sensor {
     private String deviceNum;       // 当前设备编号，格式：sen-{sensor_type_id}-{从1001开始递增}
     private Integer sensorTypeId;   // 传感器类型ID：5-温度 6-湿度 7-气体
     private String sensorName;      // 传感器名称：温度1、温度2等
-    private BigDecimal sensorValue; // 传感器值
+    private Double sensorValue;     // 传感器值（Double类型，精度足够，性能更好）
     private String unit;            // 单位：°C, %, ppm等
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
